services/agent/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.162
dbt-core==1.11.2
dbt-postgres==1.8.2
